requirements.txt
git+https://<your-github-token>@github.com/astronomer/airflow-provider-mlflow.git@main
lightgbm==3.2.1
apache-airflow-providers-slack==7.2.0
astro-sdk-python[postgres]==1.5.2
evidently==0.2.5